refactor: rename features directory
