+++ src/client/templates/Album/index.tsx
