+++ src/client/templates/Search/components/SearchResultCategory/index.tsx
